+++ test/integration/integration_tests.dart
@@ -7,6 +7,8 @@ import 'package:dart_orm_adapter_postgresql/dart_orm_adapter_postgresql.dart';
 import 'package:dart_orm_adapter_mongodb/dart_orm_adapter_mongodb.dart';
 import 'package:dart_orm_adapter_mysql/dart_orm_adapter_mysql.dart';
 
+import 'separate_annotations.dart';
+
 @ORM.DBTable('users')
 class User extends ORM.Model {
   @ORM.DBField()
@@ -163,6 +165,9 @@ allTests() {
   test('DateTime', () async {
     await dateTimeTestCase();
   });
+  test('SeparateAnnotations', () async {
+    await testSeparateAnnotations();
+  });
 }
 
 Future runIntegrationTests() async {
@@ -174,21 +179,27 @@ Future runIntegrationTests() async {
   PostgresqlDBAdapter postgresqlAdapter = new PostgresqlDBAdapter(
       'postgres://dart_orm_test:dart_orm_test@localhost:5432/dart_orm_test');
   await postgresqlAdapter.connect();
+  ORM.adapters['postgresql'] = postgresqlAdapter;
   ORM.Model.ormAdapter = postgresqlAdapter;
+  ORM.setDefaultAdapter('postgresql');
   bool migrationResult = await ORM.Migrator.migrate();
   assert(migrationResult);
 
   MongoDBAdapter mongoAdapter = new MongoDBAdapter(
       'mongodb://dart_orm_test:dart_orm_test@127.0.0.1/dart_orm_test');
   await mongoAdapter.connect();
+  ORM.adapters['mongodb'] = mongoAdapter;
   ORM.Model.ormAdapter = mongoAdapter;
+  ORM.setDefaultAdapter('mongodb');
   migrationResult = await ORM.Migrator.migrate();
   assert(migrationResult);
 
   MySQLDBAdapter mysqlAdapter = new MySQLDBAdapter(
       'mysql://dart_orm_test:dart_orm_test@localhost:3306/dart_orm_test');
   await mysqlAdapter.connect();
+  ORM.adapters['mysql'] = mysqlAdapter;
   ORM.Model.ormAdapter = mysqlAdapter;
+  ORM.setDefaultAdapter('mysql');
   migrationResult = await ORM.Migrator.migrate();
   assert(migrationResult);
 
@@ -196,6 +207,7 @@ Future runIntegrationTests() async {
     group('PostgreSQL ->', () {
       setUp(() {
         ORM.Model.ormAdapter = postgresqlAdapter;
+        ORM.setDefaultAdapter('postgresql');
       });
 
       allTests();
@@ -204,6 +216,7 @@ Future runIntegrationTests() async {
     group('MySQL ->', () {
       setUp(() {
         ORM.Model.ormAdapter = mysqlAdapter;
+        ORM.setDefaultAdapter('mysql');
       });
 
       allTests();
@@ -212,6 +225,7 @@ Future runIntegrationTests() async {
     group('MongoDB ->', () {
       setUp(() {
         ORM.Model.ormAdapter = mongoAdapter;
+        ORM.setDefaultAdapter('mongodb');
       });
 
       allTests();